keywordExtractAPI/requirements.txt
asgiref==3.5.2
Django==3.2.14
pytz==2022.1
sqlparse==0.4.2
typing_extensions==4.3.0
keras
keras_applications>=1.0.7,<=1.0.8
efficientnet
six
Translator
python-math
konlpy
pandas
djangorestframework
django-cors-headers
googletrans==3.1.0a0
protobuf==3.20.0
pillow
scikit-image
numpy
segmentation_models
sentence_transformers
scikit-learn

